gui: Get raw_current_mv and scaled_current_mv to be displayed in the status tab and in the graph window correctly.

+++ gui/qt/main_window.h
@@ -70,7 +70,7 @@ public:
   void set_up_time(uint32_t);
   void set_duty_cycle(int16_t);
   void set_raw_current_mv(uint16_t);
-  void set_scaled_current_mv(uint16_t);
+  void set_scaled_current_mv(int32_t);
   void set_current_chopping_log(uint16_t);
   void set_vin_voltage(uint16_t);
